requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
dspy-ai>=2.0.0
qdrant-client>=1.16.0
litellm>=1.0.0
mem0-ai>=0.1.0
numpy>=1.24.0
